feat(FileUpload): custom password protected PDF modal
