+++ src/components/FileUpload/components/FilePreview.styles.tsx
@@ -1,0 +1,13 @@
+import styled from 'styled-components'
+
+export const StyledFilePreview = styled.div`
+	position: absolute;
+	top: 8px;
+	left: 8px;
+	overflow: hidden;
+	width: calc(100% - 16px);
+	height: calc(100% - 16px);
+	object-fit: contain;
+	border-radius: ${({ theme }): string =>
+		theme.tokens.inputButton.borderRadius.xs};
+`
