+++ src/components/FileUpload/components/FilePreview.tsx
@@ -1,56 +1,61 @@
-import React, { useState } from 'react'
+import React from 'react'
 import { Document, Page, pdfjs } from 'react-pdf'
-import styled from 'styled-components'
 
-import { FilePreviewCommonProps, FileWithPreview } from '../index'
+import { PasswordModalQueueProps } from '../FileUpload.types'
+import { StyledFilePreview } from './FilePreview.styles'
 
 pdfjs.GlobalWorkerOptions.workerSrc = `//cdnjs.cloudflare.com/ajax/libs/pdf.js/${pdfjs.version}/pdf.worker.min.js`
 
-interface FilePreviewProps extends FilePreviewCommonProps {
+export interface FileWithPreview extends File {
+	preview: string
+}
+
+export interface FilePreviewCommonProps {
+	setFiles: (files: FileWithPreview[]) => void
+	/** Callback called when a correct password is entered for a password protected PDF document */
+	onPassword?(filename: string, password: string): void
+}
+
+interface FilePreviewProps
+	extends FilePreviewCommonProps,
+		PasswordModalQueueProps {
 	file: FileWithPreview
 }
 
-const StyledFilePreview = styled.div`
-	position: absolute;
-	top: 8px;
-	left: 8px;
-	overflow: hidden;
-	width: calc(100% - 16px);
-	height: calc(100% - 16px);
-	object-fit: contain;
-	border-radius: ${({ theme }): string =>
-		theme.tokens.inputButton.borderRadius.xs};
-`
-
 const FilePreview: React.FC<FilePreviewProps> = ({
 	file,
-	setFiles,
 	onPassword,
-	passwordPromptText,
-	passwordIncorrectText
+	passwordModalQueue,
+	setPasswordModalQueue
 }) => {
-	const [password, setPassword] = useState<string>()
-
 	const onPasswordCallback = (
 		callback: (password: string) => void,
 		reason: string
 	): void => {
-		const callbackProxy = (password: string | null): void => {
-			if (password === null) {
-				setPassword(undefined)
-				return setFiles([])
-			}
-			setPassword(password)
-			callback(password)
-		}
 		if (parseInt(reason, 10) === pdfjs.PasswordResponses.NEED_PASSWORD) {
-			const password = prompt(`${passwordPromptText} [${file.name}]`)
-			callbackProxy(password)
+			setPasswordModalQueue((queue) => [
+				...queue,
+				{
+					callback,
+					incorrect: false,
+					filename: file.name
+				}
+			])
 		} else if (
 			parseInt(reason, 10) === pdfjs.PasswordResponses.INCORRECT_PASSWORD
 		) {
-			const password = prompt(`${passwordIncorrectText} [${file.name}]`)
-			callbackProxy(password)
+			setPasswordModalQueue((queue) => {
+				const queueItem = queue.find((item) => item.filename === file.name)
+				const newQueue = queue.filter((item) => item.filename !== file.name)
+				return [
+					{
+						...queueItem,
+						callback,
+						incorrect: true
+					},
+					...newQueue
+				]
+			})
 		}
 	}
 
@@ -60,9 +65,14 @@ const FilePreview: React.FC<FilePreviewProps> = ({
 				file={file}
 				onPassword={onPasswordCallback}
 				onLoadSuccess={(): void => {
-					if (password) {
-						onPassword?.(file.name, password)
-						setPassword(undefined)
+					const currentFile = passwordModalQueue.find(
+						(item) => item.filename === file.name
+					)
+					if (currentFile?.password) {
+						onPassword?.(file.name, currentFile.password)
+						setPasswordModalQueue(
+							passwordModalQueue.filter((item) => item.filename !== file.name)
+						)
 					}
 				}}
 			>
