+++ src/demo/DemoForm/index.tsx
@@ -5,7 +5,8 @@ import { Button } from '../../components/Button'
 import { Checkbox } from '../../components/Checkbox'
 import { SelectOption } from '../../components/common/Select'
 import { DateInput, DateValue } from '../../components/DateInput'
-import { FileUpload, FileWithPreview } from '../../components/FileUpload'
+import { FileUpload } from '../../components/FileUpload'
+import { FileWithPreview } from '../../components/FileUpload/components/FilePreview'
 import { Label } from '../../components/Label'
 import { Notice } from '../../components/Notice'
 import { Select } from '../../components/Select'
